requirements.txt
uvicorn
deepgram-sdk==3.2.7
python-dotenv==1.0.0
openai==1.30.5
broadcaster==0.2.0
fastapi==0.95.0
jinja2==3.1.2
langchain==0.2.6
langchain-openai==0.1.10
langchain-core==0.2.10
langchain-community==0.2.6
beautifulsoup4==4.12.3
pymongo==4.3.3
google-api-python-client==2.137.0